requirements.txt
# Core
torch>=1.12.0
torchvision>=0.13.0

# Image processing
Pillow>=9.0.0

# Utilities
tqdm>=4.64.0
numpy>=1.23.0

# Metrics
scikit-image>=0.19.0

# Logging (choose one or both)
tensorboard>=2.9.0
wandb>=0.13.0